+++ apps/cowswap-frontend/src/common/updaters/orders/UnfillableOrdersUpdater/hooks/usePriceOutOfRangeAnalytics.ts
@@ -1,0 +1,20 @@
+import { useCallback } from 'react'
+
+import { useCowAnalytics } from '@cowprotocol/analytics'
+
+import { CowSwapAnalyticsCategory } from '../../../../analytics/types'
+
+export function usePriceOutOfRangeAnalytics(): (label: string) => void {
+  const cowAnalytics = useCowAnalytics()
+
+  return useCallback(
+    (label: string) => {
+      cowAnalytics.sendEvent({
+        category: CowSwapAnalyticsCategory.TRADE,
+        action: 'Price out of range',
+        label,
+      })
+    },
+    [cowAnalytics],
+  )
+}
