+++ apps/cowswap-frontend/src/common/updaters/orders/UnfillableOrdersUpdater/index.ts
@@ -1,0 +1,48 @@
+import { useRef } from 'react'
+
+import { useWalletInfo } from '@cowprotocol/wallet'
+
+import useSWR, { SWRConfiguration } from 'swr'
+
+import { PENDING_ORDERS_PRICE_CHECK_POLL_INTERVAL } from 'legacy/state/orders/consts'
+import { useOnlyPendingOrders } from 'legacy/state/orders/hooks'
+
+import { useUpdatePendingOrders } from './hooks/useUpdatePendingOrders'
+
+import { useIsProviderNetworkUnsupported } from '../../../hooks/useIsProviderNetworkUnsupported'
+
+const SWR_CONFIG: SWRConfiguration = {
+  refreshInterval: PENDING_ORDERS_PRICE_CHECK_POLL_INTERVAL,
+  refreshWhenHidden: false,
+  refreshWhenOffline: false,
+  revalidateOnFocus: false,
+}
+
+/**
+ * Updater that checks whether pending orders are still "fillable"
+ */
+export function UnfillableOrdersUpdater(): null {
+  const { chainId, account } = useWalletInfo()
+  const isProviderNetworkUnsupported = useIsProviderNetworkUnsupported()
+
+  const pendingOrders = useOnlyPendingOrders(chainId, account)
+  const pendingOrdersRef = useRef(pendingOrders)
+  pendingOrdersRef.current = pendingOrders
+
+  const updatePending = useUpdatePendingOrders()
+
+  const updatePendingRef = useRef(updatePending)
+  updatePendingRef.current = updatePending
+
+  useSWR(
+    account && !isProviderNetworkUnsupported && pendingOrders.length
+      ? [account, chainId, 'UnfillableOrdersUpdater']
+      : null,
+    () => {
+      return updatePendingRef.current(pendingOrdersRef.current)
+    },
+    SWR_CONFIG,
+  )
+
+  return null
+}
